6/9 Finalizing
Added new Helpers, Upgrades, Achievements.

Customize Page scrapped in favor of Stats page, which display some cool stats for the player.

I need to update spec since I changed Customize page into Stats page.

+++ IndependentProject/IndependentProject/Classes/Upgrade.cs
@@ -14,13 +14,13 @@ namespace IndependentProject.Classes
         public event PropertyChangedEventHandler PropertyChanged;
 
         public string Image { get; set; }
-        public int Cost { get; set; }
+        public long Cost { get; set; }
         public double Multiplier { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public Helper Helper { get; set; }
 
-        public Upgrade(string image, int cost, double multiplier, string name, string description, Helper helper)
+        public Upgrade(string image, long cost, double multiplier, string name, string description, Helper helper)
         {
             Image = image;
             Cost = cost;
